MNW : Refactored Unit Test Setup, Updated Opensrp Client README
